Add ordering filter to NamedModelViewSet
fixes #3266
