+++ CHANGES/3266.bugfix
@@ -1,0 +1,1 @@
+Made ordering consistent between all list API endpoints.
