+++ pulpcore/app/viewsets/content.py
@@ -4,6 +4,7 @@ from django.db import models
 from rest_framework import mixins, permissions, status
 from rest_framework.response import Response
 
+from pulpcore.filters import BaseFilterSet
 from pulpcore.app.models import Artifact, Content, PublishedMetadata, SigningService
 from pulpcore.app.serializers import (
     ArtifactSerializer,
@@ -11,7 +12,7 @@ from pulpcore.app.serializers import (
     SigningServiceSerializer,
 )
 from pulpcore.app.util import get_viewset_for_model
-from pulpcore.app.viewsets.base import BaseFilterSet, NamedModelViewSet
+from pulpcore.app.viewsets.base import NamedModelViewSet
 
 from .custom_filters import (
     ArtifactRepositoryVersionFilter,
